Fixes for Time Zones in iOS Repair and iOS Inspection apps. Removing uncalled code in O365CalendarController class in web app.

+++ iOSRepairApp/iOSRepairApp/EKNPPcommonTableView.m
@@ -84,7 +84,7 @@
             }
             else
             {
-                [self.actiondelegate showErrorMessage:[NSString stringWithFormat:@"Can't find conversation item. Error code %d,error message: %@",error.code, error.localizedDescription]];
+                [self.actiondelegate showErrorMessage:[NSString stringWithFormat:@"Can't find conversation item. Error code %ld,error message: %@",(long)error.code, error.localizedDescription]];
             }
         });
     }];
@@ -126,7 +126,7 @@
             }
             else
             {
-                [self.actiondelegate showErrorMessage:[NSString stringWithFormat:@"Can't find conversation item. Error code %ld,error message: %@",error.code, error.localizedDescription]];
+                [self.actiondelegate showErrorMessage:[NSString stringWithFormat:@"Can't find conversation item. Error code %ld,error message: %@",(long)error.code, error.localizedDescription]];
             }
         });
     }];
@@ -257,7 +257,7 @@
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath{
     
     if(self.tag == RpsConversationViewTag){
-        NSString *identifier = [NSString stringWithFormat:@"ConversationCell-%ld-%ld-%ld",tableView.tag,indexPath.section, indexPath.row];
+        NSString *identifier = [NSString stringWithFormat:@"ConversationCell-%ld-%ld-%ld",(long)tableView.tag,(long)indexPath.section, (long)indexPath.row];
         ConversationCell *cell  = [tableView dequeueReusableCellWithIdentifier:identifier];
         if (cell == nil) {
             [tableView registerNib:[UINib nibWithNibName:@"ConversationCell" bundle:nil] forCellReuseIdentifier:identifier];
@@ -270,7 +270,7 @@
         return  cell;
     }
     else if(self.tag == RpsPropertyMembersViewTag){
-        NSString *identifier = [NSString stringWithFormat:@"PropertyMembersCell-%ld-%ld-%ld",tableView.tag,indexPath.section, indexPath.row];//@"";
+        NSString *identifier = [NSString stringWithFormat:@"PropertyMembersCell-%ld-%ld-%ld",(long)tableView.tag,(long)indexPath.section, (long)indexPath.row];//@"";
         PropertyMembersCell *cell  = [tableView dequeueReusableCellWithIdentifier:identifier];
         if (cell == nil) {
             [tableView registerNib:[UINib nibWithNibName:@"PropertyMembersCell" bundle:nil] forCellReuseIdentifier:identifier];
@@ -283,7 +283,7 @@
         return  cell;
     }
     else if(self.tag == RpsOneNoteViewTag){
-        NSString *identifier = [NSString stringWithFormat:@"OneNoteCell-%ld-%ld-%ld",tableView.tag,indexPath.section, indexPath.row];//@"";
+        NSString *identifier = [NSString stringWithFormat:@"OneNoteCell-%ld-%ld-%ld",(long)tableView.tag,(long)indexPath.section, (long)indexPath.row];//@"";
         OneNoteCell *cell  = [tableView dequeueReusableCellWithIdentifier:identifier];
         if (cell == nil) {
             [tableView registerNib:[UINib nibWithNibName:@"OneNoteCell" bundle:nil] forCellReuseIdentifier:identifier];
@@ -297,7 +297,7 @@
     }
     else if(self.tag == RpsPropertyFilesViewTag
             || self.tag == RpsRecentDocumentViewTag){
-        NSString *identifier = [NSString stringWithFormat:@"PropertyFileCell-%ld-%ld-%ld",tableView.tag,indexPath.section, indexPath.row];//@"";//@"";
+        NSString *identifier = [NSString stringWithFormat:@"PropertyFileCell-%ld-%ld-%ld",(long)tableView.tag,(long)indexPath.section, (long)indexPath.row];//@"";//@"";
         PropertyFileCell *cell  = [tableView dequeueReusableCellWithIdentifier:identifier];
         if (cell == nil) {
             [tableView registerNib:[UINib nibWithNibName:@"PropertyFileCell" bundle:nil] forCellReuseIdentifier:identifier];
